requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
pydantic>=2.5.0
jinja2>=3.1.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
weasyprint>=61.0
python-docx>=1.1.0
python-pptx>=0.6.23
anthropic>=0.18.0
httpx>=0.26.0
aiofiles>=23.2.0
